Fix #36, #43, #45
Incorporates changes from #39 and #46 but actually implements a complete fix for the exclude when using regexp.

The all_vms list was being updated inside the loop that is iterating over that same list. This causes the loop to break which is why even a ```exclude=.*``` will not exclude more than one match. The fix was to move the updating of the all_vms array after the loop and only when ```found_match``` is ```true```.

This code also removes the ```all_vms.remove(vm)``` line completely from the ```save_to_config_export``` function which shouldn't be removing anything as it is for **adding** to the array, not removing.