Fix List ordering to match order
Uses ```list.insert()``` instead of ```list.append()``` so that new items are always added to the back of the array. This makes it so when the array is printed in output of this script it will be in the alphabetical order in which it was originally matches instead of reverse order. This, of course, is purely cosmetic but allows for more understandable final config output in the logs/screen.