Fix edge case where no VMs left to match
If all VMs are matched by exclude and/or vdi-export, vm-export will not match anything and thus throw an erroneous warning even though this is ok if intended by the configuration. Moving the check to an if else on the list.insert() still allows found_match to be marked true so you know your code actually is matching but just got excluded preventing the warning. This still throws a warning if it really doesn't match anything which would be the intended behaviour